requirements.txt
backcall==0.2.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gym==0.18.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mdptoolbox-hiive==4.0.3.1
mlrose-hiive==2.2.4
networkx==2.5
numpy==1.20.1
pandas==1.2.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pydot==1.4.2
pyglet==1.5.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
traitlets==5.0.5
wcwidth==0.2.5
